pyproject.toml
[project]
name = "agents-planner"
version = "0.1.0"
description = "A multi-agent system for planning and designing AI agent workflows using sequential agent architecture"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai-tools>=0.45.0",
    "google-adk>=0.5.0",
    "python-dotenv>=1.1.0",
    "typer>=0.15.4",
]
